Introduce color map class.
Introduce the class `ColorMap` which maps a range of values
in linear or logarithmic scale to a list of colors from a palette.

The methods in this class are refactored versions of the existing code
in class `ColorPalette` (palette interpolation) and `ImageProducer`
(scale mapping). The methods are removed from its original classes.

The palette drawing methods will be added to the `ImageProducer` class.