Add new bogus grid class to modify the data reading behavior.
In the original design, the operation of data loading is based on
a pixel map class, which relates each pixel in the final map image
with its nearest neighbor in the sourge grid of the layer.
The computation of the pixel map is expensive for the case of
non-regular grids in the lat-lon reference system.

This leads to a situation where the ncwms system is not usable.
In addition, this design poses a problem for producing tiled maps
of vector layers, because the boundaries of each tile does not
match properly.

Here we introduce a hack to change the data reading behavior.
Instead of computing the nearest neighbor of each pixel in the
final image, load the data of the subgrid of the original grid
that wraps the bounding box of the map. Then the position of the
entries of the subgrid can be transfomed from the original reference
system of the grid to the requested reference system of the map,
to plot the cells of the grid directly on the map image.

Since the original code relies on the pixel map class to select which
data to load from a grid variable, and since the pixel map is computed
differently depending on the actual type of the target grid, introduce
a new bogus class `MapGrid` (and `MapGridImpl`) to generate pixel maps
that imply the data reading behavior described above.

Also, introduce the package `PlotUtils` and the plot function to produce
maps in the boxfill style (other styles might be added later).
Finally, adapt the map creation in the `ImageProducer` to handle the
data from the new reading behavior and render it using the functions
in that plot package.