Switch to a more sensible hack for data loading.
The current hack for data loading will not work for vector layers
due to the broken assumptions in the `readXYComponents` method of
the `SimpleVectorLayer` class.

Switch to a less hacky approach and Adapt the PixelMap generation.

In the new approach, the class `MapGridImpl` provides:

  * the extent of the subgrid of the layer grid that wraps the map.

  * the coordinates of the vertices of the subgrid in the crs of the map.

With these changes, the `PixelMap` computation and the data extraction
are straightforward, and the `SimpleVectorLayer` implementation works
without further changes.