Adapt the map creation to new hack approach and introduce marker plots.
Adapt the map request handling and the image creation
to the new `PixelMap` hack.

Also introduce a new type of plot `MarkerPlot` with markers at given positions,
optionally rotated, scaled or colored according to the data values;
and create vector markers for the exisitng vector styles.
To have full backwards compatibility, the wind barb marker is missing.
The vector scale parameter needs to be exposed in the request parameters.