Use new data loading approach in show min/max request.
If the original approach of data loading is used,
the returned min/max values may differ from the ones
actually used to render the map.

In addition, on vector layers the returned range of magnitudes
may differ from the one computed during map rendering because
the projection is not taken into account.

Request exactly the same data used to render a map with the same request.