Switch to built-in-time list of points in map grid implementation.
Keeping a list of HorizontalPoisition elements in the map grid
implementation creates memory issues with very large datasets.
To reduce the memory consumption, store the coordinates of the points
in flat arrays, and return an instance of an abstract list
that creates the HorizontalPosition object on the fly from the
coordinates stored in the arrays.