Add subsampling of layer points before plotting.
Plotting the map using all the points in the grid of the layer
is not efficient enough for highly dense datasets.

Introduce subsampling before the plot, ignoring neighboring vertices
closer than the map resolution.

Also, tweak the PixelMap creation to avoid sorting its entries.
The order of the entries is implicitly use by the reading strategy `SCANLINE`,
and is ensured by the call to an in-house made quicksort.
There is no need for that if the entries are introduced in the right order.

Also prevent the bound checking during transformation of coordinates
in the creation of the map to save time and memory
(the check creates temporary objects because the rectilinear grids
do not cache the grid extent but create it on the fly).