Move map grid subsampling to map grid class (from plotting code).
To fix (really bad) memory consumption issues, modify the map grid class
to provide a view of the original layer grid, possibly subsampled,
instead of copying the new list of vertex coordinates.
Use an optimized approach for the case of rectilinear grids.
The creation of an already subsampled grid may also improve the data loading
efficiency when the scan line strategy is used.

Fix the grid envelop implementation to allow for the creation of empty grids,
as required by the new view behavior of the map grid class.

Take advantage of the new view behavior of the map grid in the pixel map hack,
and in the image producer.

Remove the map grid interface, and use the map grid implementation everywhere.