Fix empty range error at min/max request when there is no data in the bounding box.
The original behavior before the pixel map hack was to always load
one Float value for each pixel in the image, using null for NaN values
in the layer or for points outside the evelope of the layer.

After the pixel map hack, no value is loaded if the map bounding box
and the layer are disjoint. This causes an error when retrieving the
min/max values of the layer for such a map request.

Explicitly check for empty list of layer data values and return
a backward compatible value.