Allow reading a timeseries from an existing GridDatatype
Move the heavy part of reading a timeseries from a dataset
to a more specific function receiving an existing `GridDatatype`,
and implement the reading from a dataset using that delegated function.
This is the same approach already used for `readHorizontalPoints`,
and allows to reuse the code for the layer implementation in Thredds
(socib/thredds@timeseries_fast_read).

See Unidata/thredds#XXX.